Rename package to jnigen to avoid pakcage conflict with gdx.util Add deployment under new artifact ids so we can test this workflow
