+++ gdx-jnigen-gradle/build.gradle
@@ -2,6 +2,8 @@ plugins {
     id 'java-gradle-plugin'
 }
 
+archivesBaseName = "jnigen-runtime-gradle"
+
 dependencies {
     implementation project(":gdx-jnigen")
     implementation gradleApi()
