+++ gdx-jnigen-gradle/src/main/java/com/badlogic/gdx/jnigen/gradle/JnigenIOSJarTask.java
@@ -3,7 +3,7 @@ package com.badlogic.gdx.jnigen.gradle;
 import java.io.File;
 
 import com.badlogic.gdx.jnigen.BuildTarget;
-import com.badlogic.gdx.utils.Os;
+import com.badlogic.gdx.jnigen.Os;
 
 public class JnigenIOSJarTask extends JnigenJarTask {
 	public JnigenIOSJarTask() {
