+++ gdx-jnigen-loader/build.gradle
@@ -1,5 +1,8 @@
+archivesBaseName = "jnigen-runtime-loader"
+
 sourceCompatibility = 1.7
 targetCompatibility = 1.7
+
 
 eclipse {
     project {
