+++ gdx-jnigen-loader/src/main/java/com/badlogic/gdx/jnigen/Architecture.java
@@ -1,0 +1,19 @@
+package com.badlogic.gdx.jnigen;
+
+public enum Architecture {
+	x86, ARM, RISCV,LOONGARCH;
+
+	public String toSuffix() {
+		if (this == x86) return "";
+		else return this.name().toLowerCase();
+	}
+
+	public enum Bitness {
+		_32, _64, _128;
+
+		public String toSuffix() {
+			if (this == _32) return "";
+			else return this.name().substring(1);
+		}
+	}
+}
