+++ gdx-jnigen/build.gradle
@@ -1,3 +1,5 @@
+archivesBaseName = "jnigen-runtime"
+
 dependencies {
     implementation "com.github.javaparser:javaparser-core:3.14.14"
     api project(":gdx-jnigen-loader")
