+++ gradle.properties
@@ -1,5 +1,5 @@
 group=com.badlogicgames.gdx
-version=2.5.2
+version=2.5.3
 POM_DESCRIPTION=libGDX jnigen library
 POM_NAME=libGDX jnigen library
 POM_URL=https://github.com/libgdx/gdx-jnigen
